address UT comments
